+++ src/commands/interaction/runtime/interactions.ts
@@ -1,7 +1,7 @@
 import { AppError } from '../../../kernel/errors.ts';
 import type { ClickButton } from '../../../core/click-button.ts';
 import type { AgentDeviceRuntime, CommandContext } from '../../../runtime-contract.ts';
-import { isFillableType } from '../../../utils/snapshot-processing.ts';
+import { isFillableType } from '../../../snapshot/snapshot-processing.ts';
 import type { Point } from '../../../kernel/snapshot.ts';
 import { requireIntInRange } from '../../../utils/validation.ts';
 import { successText } from '../../../utils/success-text.ts';
