+++ src/commands/interaction/runtime/selector-read-utils.ts
@@ -1,4 +1,4 @@
-export { findNodeByLabel, resolveRefLabel } from '../../../utils/snapshot-processing.ts';
+export { findNodeByLabel, resolveRefLabel } from '../../../snapshot/snapshot-processing.ts';
 
 export function shouldScopeFind(locator: string): boolean {
   return locator === 'text' || locator === 'label' || locator === 'any';
