+++ src/daemon/screenshot-overlay.ts
@@ -9,7 +9,7 @@ import {
 import type { PNG } from '../utils/png.ts';
 import { decodePngAsync, encodePngAsync } from '../utils/png-worker-client.ts';
 import { analyzeReactNativeOverlay } from '../core/react-native-overlay.ts';
-import { findNearestAncestor, normalizeType } from '../utils/snapshot-processing.ts';
+import { findNearestAncestor, normalizeType } from '../snapshot/snapshot-processing.ts';
 import { resolveAndroidOverlaySourceRect } from './screenshot-overlay-android.ts';
 import { hasPositiveRect, rectArea, rectContains } from './screenshot-overlay-rects.ts';
