+++ src/daemon/selectors-match.ts
@@ -1,7 +1,7 @@
 import type { Platform } from '../kernel/device.ts';
 import type { SnapshotNode } from '../kernel/snapshot.ts';
 import { isNodeEditable, isNodeVisible } from '../utils/selector-node.ts';
-import { extractNodeText, normalizeType } from '../utils/snapshot-processing.ts';
+import { extractNodeText, normalizeType } from '../snapshot/snapshot-processing.ts';
 import { normalizeText } from '../utils/finders.ts';
 import type { Selector, SelectorTerm } from './selectors-parse.ts';
