+++ src/daemon/snapshot-presentation/ios/presentation.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest';
 import { attachRefs, type RawSnapshotNode } from '../../../kernel/snapshot.ts';
-import { buildSnapshotVisibility } from '../../../utils/snapshot-visibility.ts';
+import { buildSnapshotVisibility } from '../../../snapshot/snapshot-visibility.ts';
 import { presentIosInteractiveSnapshot } from './index.ts';
 
 function buildSnapshotState(data: { nodes?: RawSnapshotNode[]; backend?: 'xctest' }) {
