+++ src/snapshot/snapshot-label-signals.ts
