+++ src/snapshot/snapshot-visibility.ts
